+++ lib/slimmer/processors/body_inserter.rb
@@ -10,6 +10,8 @@ module Slimmer::Processors
       source_markup = src.at_css(@source_selector)
       destination_markup = dest.at_css(@destination_selector)
 
+      raise(Slimmer::SourceWrapperNotFoundError, "Source wrapper div not found", caller) if source_markup.nil? && wrapper_check?
+
       css_classes = []
       css_classes << source_markup.attributes["class"].to_s.split(/ +/) if source_markup.has_attribute?("class")
       css_classes << destination_markup.attributes["class"].to_s.split(/ +/) if destination_markup.has_attribute?("class")
@@ -23,6 +25,10 @@ module Slimmer::Processors
 
     def is_gem_layout?
       @headers[Slimmer::Headers::TEMPLATE_HEADER]&.start_with?("gem_layout")
+    end
+
+    def wrapper_check?
+      ENV["SLIMMER_WRAPPER_CHECK"] == "true"
     end
   end
 end